refactor: i18n tests cleanup
